[+] Split server init in different file
