+++ src/zappy_server_src/core/server_data.c
@@ -11,30 +11,6 @@
 #include "argument_handling.h"
 #include <stdlib.h>
 
-server_data_t *init_server_data(int ac, char **av)
-{
-    server_data_t *server_data = malloc(sizeof(server_data_t) * 1);
-    int size_queue = 0;
-
-    if (server_data == NULL)
-        return NULL;
-    if (!(server_data->arguments = argument_handling(ac, av)))
-        return NULL;
-    server_data->active_players = malloc(sizeof(player_list_t *) * 1);
-    if (server_data->active_players == NULL)
-        return NULL;
-    server_data->active_players[0] = NULL;
-    server_data->active_player_n = 0;
-    for (int i = 0; server_data->arguments->team_list[i]; i++)
-        size_queue += server_data->arguments->client_nb;
-    server_data->server = create_new_server(server_data->arguments->port,
-    size_queue);
-    if (server_data->server == NULL)
-        return NULL;
-    server_data->server->state = true;
-    return server_data;
-}
-
 bool server_add_player(server_data_t *server_data)
 {
     player_list_t *new_player = malloc(sizeof(player_list_t) * 1);
